+++ test/resolve/forceIgnore.test.ts
@@ -9,11 +9,7 @@ import { expect } from 'chai';
 import { createSandbox } from 'sinon';
 import * as fs from 'fs';
 import * as fsUtil from '../../src/utils/fileSystemHandler';
-import { dirname, join, relative } from 'path';
-import { Lifecycle } from '@salesforce/core';
-// @ts-ignore this doesn't have typings
-import * as gitignoreParser from 'gitignore-parser';
-import ignore from 'ignore';
+import { join } from 'path';
 const env = createSandbox();
 
 describe('ForceIgnore', () => {
@@ -22,28 +18,6 @@ describe('ForceIgnore', () => {
   const testPattern = '**/__tests__/**';
 
   afterEach(() => env.restore());
-
-  it('should emit user warning once per force ignore difference', () => {
-    const readStub = env.stub(fs, 'readFileSync');
-    readStub.withArgs(forceIgnorePath).returns('*__tests__*');
-    // @ts-ignore private member
-    const emitWarningSpy = env.stub(ForceIgnore.prototype, 'emitWarning');
-    const forceIgnore = new ForceIgnore(forceIgnorePath);
-    const files = [
-      join('some', 'path', '__tests__', 'myTest.x'),
-      join('some', 'path', 'other', '__tests__', 'myOtherTest.x'),
-    ];
-    // @ts-ignore private member
-    env.stub(forceIgnore.gitignoreParser, 'accepts').returns(false);
-    // @ts-ignore private member
-    env.stub(forceIgnore.parser, 'ignores').returns(false);
-    // @ts-ignore parseContents is private
-    env.stub(gitignoreParser, 'compile').returns({
-      accepts: () => true,
-    });
-    files.forEach((file) => forceIgnore.accepts(file));
-    expect(emitWarningSpy.callCount).to.be.equal(2);
-  });
 
   it('Should default to not ignoring a file if forceignore is not loaded', () => {
     const path = join('some', 'path');
@@ -67,81 +41,6 @@ describe('ForceIgnore', () => {
     searchStub.withArgs(testPath, ForceIgnore.FILE_NAME).returns(forceIgnorePath);
     const forceIgnore = ForceIgnore.findAndCreate(testPath);
     expect(forceIgnore.accepts(testPath)).to.be.false;
-  });
-
-  it('should send telemetry event when the old parser and the new parser have different results testing old parser', () => {
-    const readStub = env.stub(fs, 'readFileSync');
-    const testPattern = '*__tests__*';
-    readStub.withArgs(forceIgnorePath).returns(testPattern);
-    const forceIgnore = new ForceIgnore(forceIgnorePath);
-    const file = join('some', 'path', '__tests__', 'myTest.x');
-    // @ts-ignore private member
-    env.stub(forceIgnore.gitignoreParser, 'accepts').returns(false);
-    // @ts-ignore private member
-    env.stub(forceIgnore.parser, 'ignores').returns(false);
-
-    // @ts-ignore parseContents is private
-    const oldParser = gitignoreParser.compile(forceIgnore.parseContents(testPattern));
-    const newParser = ignore().add([testPattern]);
-
-    env.stub(oldParser, 'accepts').returns(false);
-    env.stub(newParser, 'ignores').returns(true);
-
-    const telemetrySpy = env.stub(Lifecycle.prototype, 'emit');
-
-    const expected = {
-      eventName: 'FORCE_IGNORE_DIFFERENCE',
-      content: testPattern,
-      oldLibraryResults: false,
-      newLibraryResults: true,
-      ignoreLines: [testPattern],
-      file: relative(dirname(forceIgnorePath), file),
-    };
-
-    forceIgnore.accepts(file);
-
-    expect(telemetrySpy.calledOnce).to.be.true;
-    expect(telemetrySpy.args[0][0]).to.equal('telemetry');
-    expect(telemetrySpy.args[0][1]).to.deep.equal(expected);
-  });
-
-  it('should send telemetry event when the old parser and the new parser have different results testing new parser', () => {
-    const readStub = env.stub(fs, 'readFileSync');
-    const switchParser = '# .forceignore v2';
-    const testPattern = '*__tests__*';
-    const entries = testPattern + '\n' + switchParser;
-    readStub.withArgs(forceIgnorePath).returns(entries);
-    const forceIgnore = new ForceIgnore(forceIgnorePath);
-    const telemetrySpy = env.spy(Lifecycle.prototype, 'emit');
-
-    const file = join('some', 'path', '__tests__', 'myTest.x');
-
-    // @ts-ignore private member
-    env.stub(forceIgnore.gitignoreParser, 'accepts').returns(false);
-    // @ts-ignore private member
-    env.stub(forceIgnore.parser, 'ignores').returns(false);
-
-    // @ts-ignore parseContents is private
-    const oldParser = gitignoreParser.compile(forceIgnore.parseContents(testPattern));
-    const newParser = ignore().add([testPattern]);
-
-    env.stub(oldParser, 'accepts').returns(false);
-    env.stub(newParser, 'ignores').returns(true);
-
-    const expected = {
-      eventName: 'FORCE_IGNORE_DIFFERENCE',
-      content: entries,
-      oldLibraryResults: false,
-      newLibraryResults: true,
-      ignoreLines: [testPattern, switchParser],
-      file: relative(dirname(forceIgnorePath), file),
-    };
-
-    forceIgnore.accepts(file);
-
-    expect(telemetrySpy.calledOnce).to.be.true;
-    expect(telemetrySpy.args[0][0]).to.equal('telemetry');
-    expect(telemetrySpy.args[0][1]).to.deep.equal(expected);
   });
 
   it('Should handle forward slashes on windows', () => {
@@ -169,15 +68,12 @@ describe('ForceIgnore', () => {
     const root = join('some', 'path');
 
     beforeEach(() => {
-      // when using the new parser, respect the defaults
-      env.stub(fs, 'readFileSync').returns('# .forceignore v2');
+      env.stub(fs, 'readFileSync').returns('');
       forceIgnore = new ForceIgnore();
     });
 
     it('Should ignore files starting with a dot', () => {
       const dotPath = join(root, '.xyz');
-      const telemetrySpy = env.spy(Lifecycle.prototype, 'emit');
-      expect(telemetrySpy.called).to.be.false;
 
       expect(forceIgnore.accepts(dotPath)).to.be.false;
       expect(forceIgnore.denies(dotPath)).to.be.true;
@@ -185,8 +81,6 @@ describe('ForceIgnore', () => {
 
     it('Should ignore files ending in .dup', () => {
       const dupPath = join(root, 'abc.dup');
-      const telemetrySpy = env.spy(Lifecycle.prototype, 'emit');
-      expect(telemetrySpy.called).to.be.false;
 
       expect(forceIgnore.accepts(dupPath)).to.be.false;
       expect(forceIgnore.denies(dupPath)).to.be.true;
@@ -202,43 +96,6 @@ describe('ForceIgnore', () => {
       const manifestPath = join(root, 'package2-manifest.json');
       expect(forceIgnore.accepts(manifestPath)).to.be.false;
       expect(forceIgnore.denies(manifestPath)).to.be.true;
-    });
-  });
-
-  describe('Defaults with old parser', () => {
-    let forceIgnore: ForceIgnore;
-    const root = join('some', 'path');
-
-    beforeEach(() => {
-      env.stub(fs, 'readFileSync').returns('');
-      forceIgnore = new ForceIgnore();
-    });
-
-    it('Should ignore files starting with a dot', () => {
-      const dotPath = join(root, '.xyz');
-      expect(forceIgnore.accepts(dotPath)).to.be.true;
-      expect(forceIgnore.denies(dotPath)).to.be.false;
-    });
-
-    it('Should ignore files ending in .dup', () => {
-      const dupPath = join(root, 'abc.dup');
-      const telemetrySpy = env.spy(Lifecycle.prototype, 'emit');
-      expect(telemetrySpy.called).to.be.false;
-
-      expect(forceIgnore.accepts(dupPath)).to.be.true;
-      expect(forceIgnore.denies(dupPath)).to.be.false;
-    });
-
-    it('Should ignore files named package2-descriptor.json', () => {
-      const descriptorPath = join(root, 'package2-descriptor.json');
-      expect(forceIgnore.accepts(descriptorPath)).to.be.true;
-      expect(forceIgnore.denies(descriptorPath)).to.be.false;
-    });
-
-    it('Should ignore files named package2-manifest.json', () => {
-      const manifestPath = join(root, 'package2-manifest.json');
-      expect(forceIgnore.accepts(manifestPath)).to.be.true;
-      expect(forceIgnore.denies(manifestPath)).to.be.false;
     });
   });
 });